requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
Flask-Session==0.8.0
gunicorn==23.0.0
python-dotenv==1.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
requests==2.32.5
pdfkit==1.0.0
